requirements.txt
deepface>=0.0.75
retina-face>=0.0.12
matplotlib>=3.6.2
opencv-python>=4.7.0.71
streamlit>=1.28.0
insightface>=0.7.0
onnxruntime>=1.15.0
scikit-learn>=1.3.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
pandas>=2.0.0
